+++ .circleci/config.yml
@@ -488,8 +488,8 @@ workflows:
           extras: '[redshift-binary,itest]'
           python_version: "3.6"
           db_name: demo-itest
-          requires:
-            - test-3.8
+#          requires:  # TODO
+#            - test-3.8
           filters:
             tags:
               only: /v\d+\.\d+\.\d+(-[\w]+)?/
