+++ tests/unit/url/test_url_defaults.py
@@ -1,14 +1,16 @@
-from records_mover.url import init_urls
-from records_mover import url
+from records_mover.url.resolver import init_urls
+from records_mover.url import resolver
 import unittest
 
 
 class TestUrl(unittest.TestCase):
     def setUp(self):
-        url.file_url_ctors = {}
-        url.directory_url_ctors = {}
+        resolver.file_url_ctors = {}
+        resolver.directory_url_ctors = {}
 
     def test_init_urls(self):
         init_urls()
-        self.assertEqual(list(url.directory_url_ctors.keys()), ['s3', 'gs', 'file'])
-        self.assertEqual(list(url.file_url_ctors.keys()), ['s3', 'gs', 'file', 'http', 'https'])
+        self.assertEqual(list(resolver.directory_url_ctors.keys()),
+                         ['s3', 'gs', 'file'])
+        self.assertEqual(list(resolver.file_url_ctors.keys()),
+                         ['s3', 'gs', 'file', 'http', 'https'])
